refactor(account): 精简 Hub 设置列表——视觉瘦身 + 折叠应用设置为子页
个人中心下方设置项过多、每行副标题让列表显长。两处优化：

A 视觉瘦身：去掉自明行的副标题（外观改为右侧显示当前主题值、我的收藏、
  邮件订阅、每日精选提醒、反馈、关于我们），分组从 4 组并到 3 组。
C 长尾折叠：把低频工具性偏好（应用语言/摘要语言/默认首页/外链打开方式）从
  Hub 折叠进新的「应用设置」子页 AppSettingsScreen，Hub 上只留一行入口
  （通用组，带「语言、默认首页、打开方式」副标题 + 右箭头）。摘要语言的意图
  采集 + 赞助流程随之迁入子页。

Hub 设置列表由「4 组 10 行（多数带副标题）」缩为「3 组 7 行（多为单行）」，
几乎一屏放下。清理 AccountScreen 46 个随之失效的 import。

验证：Android+iOS 编译通过、shared 单测全绿；Pixel_9_2 冒烟——瘦身后的 Hub
与「应用设置」子页渲染正确、折叠入口跳转与返回导航正常。
