+++ shared/src/commonMain/composeResources/values-zh/strings.xml
@@ -193,6 +193,8 @@
     <string name="account_pro_active">你已是 Pro 会员，感谢支持！</string>
     <string name="account_github_entry">GitHub 主页</string>
     <string name="account_github_entry_desc">贡献与关注动态</string>
+    <string name="settings_group_general">通用</string>
+    <string name="app_settings_entry_desc">语言、默认首页、打开方式</string>
     <string name="contrib_total">最近一年 %1$s 次贡献</string>
     <string name="contrib_less">少</string>
     <string name="contrib_more">多</string>
