+++ shared/src/commonMain/kotlin/whl/trending/ai/ui/profile/AccountScreen.kt
@@ -22,17 +22,12 @@ import androidx.compose.material.icons.filled.AccountCircle
 import androidx.compose.material.icons.filled.Email
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.Feedback
-import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Info
-import androidx.compose.material.icons.filled.Language
 import androidx.compose.material.icons.filled.Notifications
-import androidx.compose.material.icons.filled.OpenInBrowser
 import androidx.compose.material.icons.filled.Palette
-import androidx.compose.material.icons.filled.Translate
+import androidx.compose.material.icons.filled.Tune
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.FilledTonalButton
@@ -45,7 +40,6 @@ import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.LoadingIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedCard
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
@@ -75,9 +69,7 @@ import coil3.compose.AsyncImage
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.stringResource
 import trendingai.shared.generated.resources.Res
-import trendingai.shared.generated.resources.about
 import trendingai.shared.generated.resources.about_us
-import trendingai.shared.generated.resources.about_us_desc
 import trendingai.shared.generated.resources.account_github_entry
 import trendingai.shared.generated.resources.account_github_entry_desc
 import trendingai.shared.generated.resources.account_plan_title
@@ -89,35 +81,17 @@ import trendingai.shared.generated.resources.account_tier_free
 import trendingai.shared.generated.resources.account_title
 import trendingai.shared.generated.resources.account_upgrade_cta
 import trendingai.shared.generated.resources.account_upgrade_hint
-import trendingai.shared.generated.resources.app_language_desc
 import trendingai.shared.generated.resources.app_settings
+import trendingai.shared.generated.resources.app_settings_entry_desc
 import trendingai.shared.generated.resources.appearance
 import trendingai.shared.generated.resources.back
 import trendingai.shared.generated.resources.cancel
-import trendingai.shared.generated.resources.close
 import trendingai.shared.generated.resources.daily_picks_notification
-import trendingai.shared.generated.resources.daily_picks_notification_desc
-import trendingai.shared.generated.resources.default_home_tab
-import trendingai.shared.generated.resources.default_home_tab_desc
 import trendingai.shared.generated.resources.favorites
 import trendingai.shared.generated.resources.feedback
-import trendingai.shared.generated.resources.feedback_desc
-import trendingai.shared.generated.resources.feedback_email_invalid
-import trendingai.shared.generated.resources.feedback_email_placeholder
-import trendingai.shared.generated.resources.feedback_error
-import trendingai.shared.generated.resources.feedback_rate_limit
-import trendingai.shared.generated.resources.hackernews_title
-import trendingai.shared.generated.resources.language_option_chinese
-import trendingai.shared.generated.resources.language_option_english
-import trendingai.shared.generated.resources.language_option_follow_system
-import trendingai.shared.generated.resources.language_settings
 import trendingai.shared.generated.resources.notification_permission_denied
-import trendingai.shared.generated.resources.open_links_in_browser
-import trendingai.shared.generated.resources.open_links_in_browser_desc
 import trendingai.shared.generated.resources.open_system_settings
 import trendingai.shared.generated.resources.personalization
-import trendingai.shared.generated.resources.picks_title
-import trendingai.shared.generated.resources.producthunt_title
 import trendingai.shared.generated.resources.profile_load_failed
 import trendingai.shared.generated.resources.profile_quota_error
 import trendingai.shared.generated.resources.profile_quota_rates
@@ -128,47 +102,29 @@ import trendingai.shared.generated.resources.profile_followers
 import trendingai.shared.generated.resources.profile_repos
 import trendingai.shared.generated.resources.profile_retry
 import trendingai.shared.generated.resources.settings
+import trendingai.shared.generated.resources.settings_group_general
 import trendingai.shared.generated.resources.sign_in
 import trendingai.shared.generated.resources.sign_out
 import trendingai.shared.generated.resources.sign_out_confirm
-import trendingai.shared.generated.resources.subscribe_desc
 import trendingai.shared.generated.resources.subscribe_title
 import trendingai.shared.generated.resources.subscription_reminders
-import trendingai.shared.generated.resources.summary_lang_capture_identity_note
-import trendingai.shared.generated.resources.summary_lang_capture_label
-import trendingai.shared.generated.resources.summary_lang_capture_message
-import trendingai.shared.generated.resources.summary_lang_capture_title
-import trendingai.shared.generated.resources.summary_language
-import trendingai.shared.generated.resources.summary_language_desc
-import trendingai.shared.generated.resources.summary_language_feedback
-import trendingai.shared.generated.resources.summary_language_message
-import trendingai.shared.generated.resources.summary_language_sponsor
-import whl.trending.ai.auth.AuthState
 import whl.trending.ai.auth.globalAuthManager
 import whl.trending.ai.core.DateTimeUtils
 import whl.trending.ai.core.ProSponsor
-import whl.trending.ai.core.isValidEmail
-import whl.trending.ai.core.platform.getSystemLanguage
-import whl.trending.ai.core.platform.getSystemLanguageDisplayName
-import whl.trending.ai.core.platform.isIosPlatform
 import whl.trending.ai.core.platform.openAppSettings
 import whl.trending.ai.core.platform.trackEvent
-import whl.trending.ai.data.local.AppLanguage
-import whl.trending.ai.data.local.SummaryLanguage
 import whl.trending.ai.data.local.ThemeMode
 import whl.trending.ai.data.local.globalSettingsManager
 import whl.trending.ai.data.model.QuotaResponse
-import whl.trending.ai.data.remote.ApiException
-import whl.trending.ai.data.repository.TrendingRepository
 import whl.trending.ai.notification.globalDailyPicksNotifier
-import whl.trending.ai.ui.home.HomeTab
 import whl.trending.ai.ui.settings.themeModeText
 
 /**
  * 账户中心（统一 Hub）：应用「关于我」的唯一入口，替代原独立的个人主页 + 设置页。
  *
  * 从上到下：身份区（含未登录引导）→ 套餐 & 用量（含升级 CTA）→ GitHub 主页入口卡
- * （仅 GitHub 用户）→ 设置分组（个性化 / 订阅提醒 / 应用 / 关于）→ 退出登录。
+ * （仅 GitHub 用户）→ 设置分组（个性化 / 订阅提醒 / 通用）→ 退出登录。
+ * 低频应用偏好（语言/默认首页/打开方式）折叠进「应用设置」子页 [AppSettingsScreen]。
  *
  * 未登录用户同样可达：身份区显示登录引导、用量区显示匿名额度、设置分组照常可用——
  * 不再像旧个人主页那样对邮箱/匿名用户只剩空壳。GitHub 开发者档案（贡献图 + 动态流）
@@ -183,6 +139,7 @@ fun AccountScreen(
     onNavigateToFeedback: () -> Unit = {},
     onNavigateToSubscribe: () -> Unit = {},
     onNavigateToAppearance: () -> Unit = {},
+    onNavigateToAppSettings: () -> Unit = {},
     onNavigateToAbout: () -> Unit = {},
 ) {
     val viewModel: ProfileViewModel = viewModel { ProfileViewModel() }
@@ -192,26 +149,14 @@ fun AccountScreen(
     val authSupported = globalAuthManager.isSupported
     LaunchedEffect(Unit) { if (authSupported) viewModel.load() }
 
-    val isIos = isIosPlatform()
-    val authState by globalAuthManager.authState.collectAsState()
-    val isLoggedIn = authState is AuthState.LoggedIn
     val isPro by globalSettingsManager.isPro.collectAsState(
         initial = globalSettingsManager.currentIsPro()
     )
-
-    val appLanguage by globalSettingsManager.appLanguage.collectAsState(AppLanguage.FOLLOW_SYSTEM)
-    val summaryLanguage by globalSettingsManager.summaryLanguage.collectAsState(SummaryLanguage.FOLLOW_SYSTEM)
     val themeMode by globalSettingsManager.themeMode.collectAsState(ThemeMode.FOLLOW_SYSTEM)
-    val openLinksInCustomTab by globalSettingsManager.openLinksInCustomTab.collectAsState(true)
-    val defaultHomeTab by globalSettingsManager.defaultHomeTab.collectAsState(
-        remember { globalSettingsManager.currentDefaultHomeTab() }
-    )
     val dailyPicksNotificationEnabled by globalSettingsManager.dailyPicksNotificationEnabled.collectAsState(
         remember { globalSettingsManager.currentDailyPicksNotificationEnabled() }
     )
 
-    var showSummaryLanguageDialog by remember { mutableStateOf(false) }
-    var showLangCaptureDialog by remember { mutableStateOf(false) }
     var showSignOutDialog by remember { mutableStateOf(false) }
 
     val snackbarHostState = remember { SnackbarHostState() }
@@ -219,40 +164,6 @@ fun AccountScreen(
     val permissionDeniedMsg = stringResource(Res.string.notification_permission_denied)
     val openSystemSettingsLabel = stringResource(Res.string.open_system_settings)
     val pullToRefreshState = rememberPullToRefreshState()
-
-    if (showSummaryLanguageDialog) {
-        AlertDialog(
-            onDismissRequest = { showSummaryLanguageDialog = false },
-            title = { Text(stringResource(Res.string.summary_language)) },
-            text = { Text(stringResource(Res.string.summary_language_message)) },
-            confirmButton = {
-                TextButton(onClick = {
-                    showSummaryLanguageDialog = false
-                    showLangCaptureDialog = true
-                }) {
-                    Text(stringResource(Res.string.summary_language_sponsor))
-                }
-            },
-            dismissButton = {
-                Row {
-                    TextButton(onClick = { showSummaryLanguageDialog = false }) {
-                        Text(stringResource(Res.string.close))
-                    }
-                    TextButton(onClick = {
-                        showSummaryLanguageDialog = false
-                        trackEvent("settings_summary_language_feedback")
-                        onNavigateToFeedback()
-                    }) {
-                        Text(stringResource(Res.string.summary_language_feedback))
-                    }
-                }
-            }
-        )
-    }
-
-    if (showLangCaptureDialog) {
-        LanguageCaptureDialog(isLoggedIn = isLoggedIn, onDismiss = { showLangCaptureDialog = false })
-    }
 
     if (showSignOutDialog) {
         AlertDialog(
@@ -345,8 +256,13 @@ fun AccountScreen(
                 item(key = "appearance") {
                     ListItem(
                         headlineContent = { Text(stringResource(Res.string.appearance)) },
-                        supportingContent = { Text(themeModeText(themeMode)) },
                         leadingContent = { Icon(Icons.Default.Palette, null) },
+                        trailingContent = {
+                            Text(
+                                text = themeModeText(themeMode),
+                                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                            )
+                        },
                         modifier = Modifier.clickable {
                             trackEvent("settings_appearance")
                             onNavigateToAppearance()
@@ -370,7 +286,6 @@ fun AccountScreen(
                 item(key = "subscribe") {
                     ListItem(
                         headlineContent = { Text(stringResource(Res.string.subscribe_title)) },
-                        supportingContent = { Text(stringResource(Res.string.subscribe_desc)) },
                         leadingContent = { Icon(Icons.Default.Email, null) },
                         modifier = Modifier.clickable {
                             trackEvent("settings_subscribe")
@@ -382,7 +297,6 @@ fun AccountScreen(
                     item(key = "daily_picks_notification") {
                         ListItem(
                             headlineContent = { Text(stringResource(Res.string.daily_picks_notification)) },
-                            supportingContent = { Text(stringResource(Res.string.daily_picks_notification_desc)) },
                             leadingContent = { Icon(Icons.Default.Notifications, null) },
                             trailingContent = {
                                 Switch(
@@ -418,145 +332,25 @@ fun AccountScreen(
                 }
                 item(key = "divider_2") { HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp)) }
 
-                // 应用设置
-                item(key = "group_app") { SettingsHeader(stringResource(Res.string.app_settings)) }
-                item(key = "app_language") {
-                    var expanded by remember { mutableStateOf(false) }
+                // 通用：应用设置（低频偏好折叠进子页）+ 反馈 + 关于
+                item(key = "group_general") { SettingsHeader(stringResource(Res.string.settings_group_general)) }
+                item(key = "app_settings_entry") {
                     ListItem(
-                        headlineContent = { Text(stringResource(Res.string.language_settings)) },
-                        supportingContent = { Text(stringResource(Res.string.app_language_desc)) },
-                        leadingContent = { Icon(Icons.Default.Language, null) },
+                        headlineContent = { Text(stringResource(Res.string.app_settings)) },
+                        supportingContent = { Text(stringResource(Res.string.app_settings_entry_desc)) },
+                        leadingContent = { Icon(Icons.Default.Tune, null) },
                         trailingContent = {
-                            if (isIos) {
-                                Text(
-                                    text = stringResource(Res.string.open_system_settings),
-                                    color = MaterialTheme.colorScheme.primary,
-                                    modifier = Modifier.clickable { openAppSettings() }
-                                )
-                            } else {
-                                Box {
-                                    Text(
-                                        text = languageOptionText(appLanguage),
-                                        color = MaterialTheme.colorScheme.primary,
-                                        modifier = Modifier.clickable { expanded = true }
-                                    )
-                                    DropdownMenu(
-                                        expanded = expanded,
-                                        onDismissRequest = { expanded = false }
-                                    ) {
-                                        AppLanguage.entries.forEach { language ->
-                                            DropdownMenuItem(
-                                                text = { Text(languageOptionText(language)) },
-                                                onClick = {
-                                                    expanded = false
-                                                    trackEvent("settings_language_change", mapOf("language" to language.name.lowercase()))
-                                                    globalSettingsManager.setLanguage(language)
-                                                }
-                                            )
-                                        }
-                                    }
-                                }
-                            }
+                            Icon(Icons.AutoMirrored.Filled.KeyboardArrowRight, contentDescription = null)
                         },
                         modifier = Modifier.clickable {
-                            if (isIos) openAppSettings() else expanded = true
+                            trackEvent("settings_app_settings")
+                            onNavigateToAppSettings()
                         }
                     )
                 }
-                item(key = "summary_language") {
-                    var expanded by remember { mutableStateOf(false) }
-                    ListItem(
-                        headlineContent = { Text(stringResource(Res.string.summary_language)) },
-                        supportingContent = { Text(stringResource(Res.string.summary_language_desc)) },
-                        leadingContent = { Icon(Icons.Default.Translate, null) },
-                        trailingContent = {
-                            Box {
-                                Text(
-                                    text = languageOptionText(summaryLanguage),
-                                    color = MaterialTheme.colorScheme.primary,
-                                    modifier = Modifier.clickable { expanded = true }
-                                )
-                                DropdownMenu(
-                                    expanded = expanded,
-                                    onDismissRequest = { expanded = false }
-                                ) {
-                                    SummaryLanguage.entries.forEach { language ->
-                                        DropdownMenuItem(
-                                            text = { Text(languageOptionText(language)) },
-                                            onClick = {
-                                                expanded = false
-                                                trackEvent("settings_summary_language_change", mapOf("language" to language.name.lowercase()))
-                                                globalSettingsManager.setSummaryLanguage(language)
-                                            }
-                                        )
-                                    }
-                                }
-                            }
-                        },
-                        modifier = Modifier.clickable {
-                            trackEvent("settings_summary_language", mapOf("summary_language" to summaryLanguage.name.lowercase()))
-                            showSummaryLanguageDialog = true
-                        }
-                    )
-                }
-                item(key = "default_home_tab") {
-                    var expanded by remember { mutableStateOf(false) }
-                    ListItem(
-                        headlineContent = { Text(stringResource(Res.string.default_home_tab)) },
-                        supportingContent = { Text(stringResource(Res.string.default_home_tab_desc)) },
-                        leadingContent = { Icon(Icons.Default.Home, null) },
-                        trailingContent = {
-                            Box {
-                                Text(
-                                    text = homeTabOptionText(HomeTab.fromNameOrDefault(defaultHomeTab)),
-                                    color = MaterialTheme.colorScheme.primary,
-                                    modifier = Modifier.clickable { expanded = true }
-                                )
-                                DropdownMenu(
-                                    expanded = expanded,
-                                    onDismissRequest = { expanded = false }
-                                ) {
-                                    HomeTab.entries.forEach { tab ->
-                                        DropdownMenuItem(
-                                            text = { Text(homeTabOptionText(tab)) },
-                                            onClick = {
-                                                expanded = false
-                                                trackEvent("settings_default_home_tab_change", mapOf("tab" to tab.name.lowercase()))
-                                                globalSettingsManager.setDefaultHomeTab(tab.name)
-                                            }
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                    )
-                }
-                item(key = "open_links") {
-                    val toggleOpenLinks = { enabled: Boolean ->
-                        trackEvent("settings_open_links_in_browser", mapOf("enabled" to enabled.toString()))
-                        globalSettingsManager.setOpenLinksInCustomTab(enabled)
-                    }
-                    ListItem(
-                        headlineContent = { Text(stringResource(Res.string.open_links_in_browser)) },
-                        supportingContent = { Text(stringResource(Res.string.open_links_in_browser_desc)) },
-                        leadingContent = { Icon(Icons.Default.OpenInBrowser, null) },
-                        trailingContent = {
-                            Switch(
-                                checked = openLinksInCustomTab,
-                                onCheckedChange = toggleOpenLinks
-                            )
-                        },
-                        modifier = Modifier.clickable { toggleOpenLinks(!openLinksInCustomTab) }
-                    )
-                }
-                item(key = "divider_3") { HorizontalDivider(modifier = Modifier.padding(vertical = 8.dp)) }
-
-                // 关于
-                item(key = "group_about") { SettingsHeader(stringResource(Res.string.about)) }
                 item(key = "feedback") {
                     ListItem(
                         headlineContent = { Text(stringResource(Res.string.feedback)) },
-                        supportingContent = { Text(stringResource(Res.string.feedback_desc)) },
                         leadingContent = { Icon(Icons.Default.Feedback, null) },
                         modifier = Modifier.clickable {
                             trackEvent("settings_feedback")
@@ -567,7 +361,6 @@ fun AccountScreen(
                 item(key = "about_us") {
                     ListItem(
                         headlineContent = { Text(stringResource(Res.string.about_us)) },
-                        supportingContent = { Text(stringResource(Res.string.about_us_desc)) },
                         leadingContent = { Icon(Icons.Default.Info, null) },
                         modifier = Modifier.clickable {
                             trackEvent("settings_about")
@@ -863,148 +656,5 @@ fun SettingsHeader(text: String) {
         style = MaterialTheme.typography.labelLarge,
         color = MaterialTheme.colorScheme.primary,
         modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-    )
-}
-
-@Composable
-private fun languageOptionText(language: AppLanguage): String {
-    val labelRes = when (language) {
-        AppLanguage.FOLLOW_SYSTEM -> Res.string.language_option_follow_system
-        AppLanguage.CHINESE -> Res.string.language_option_chinese
-        AppLanguage.ENGLISH -> Res.string.language_option_english
-    }
-    return stringResource(labelRes)
-}
-
-@Composable
-private fun languageOptionText(language: SummaryLanguage): String {
-    val labelRes = when (language) {
-        SummaryLanguage.FOLLOW_SYSTEM -> Res.string.language_option_follow_system
-        SummaryLanguage.CHINESE -> Res.string.language_option_chinese
-        SummaryLanguage.ENGLISH -> Res.string.language_option_english
-    }
-    return stringResource(labelRes)
-}
-
-@Composable
-private fun homeTabOptionText(tab: HomeTab): String = when (tab) {
-    HomeTab.GitHub -> "GitHub"
-    HomeTab.HackerNews -> stringResource(Res.string.hackernews_title)
-    HomeTab.ProductHunt -> stringResource(Res.string.producthunt_title)
-    HomeTab.Picks -> stringResource(Res.string.picks_title)
-}
-
-/**
- * 摘要语言意图采集弹窗：期望语言（+ 未登录的选填邮箱）经反馈接口提交，成功后跳赞助页。
- * 状态自持有——挂载即全新、关闭即丢弃。
- */
-@OptIn(ExperimentalMaterial3ExpressiveApi::class)
-@Composable
-private fun LanguageCaptureDialog(isLoggedIn: Boolean, onDismiss: () -> Unit) {
-    var langInput by remember { mutableStateOf(getSystemLanguageDisplayName()) }
-    var langEmail by remember { mutableStateOf("") }
-    var langEmailInvalid by remember { mutableStateOf(false) }
-    var langSubmitting by remember { mutableStateOf(false) }
-    var langError by remember { mutableStateOf<String?>(null) }
-    val scope = rememberCoroutineScope()
-    val errGeneric = stringResource(Res.string.feedback_error)
-    val errRate = stringResource(Res.string.feedback_rate_limit)
-
-    AlertDialog(
-        onDismissRequest = { if (!langSubmitting) onDismiss() },
-        title = { Text(stringResource(Res.string.summary_lang_capture_title)) },
-        text = {
-            Column {
-                Text(stringResource(Res.string.summary_lang_capture_message))
-                Spacer(Modifier.height(16.dp))
-                OutlinedTextField(
-                    value = langInput,
-                    onValueChange = { langInput = it; langError = null },
-                    label = { Text(stringResource(Res.string.summary_lang_capture_label)) },
-                    singleLine = true,
-                    isError = langError != null,
-                    enabled = !langSubmitting,
-                    supportingText = langError?.let { { Text(it, color = MaterialTheme.colorScheme.error) } },
-                    modifier = Modifier.fillMaxWidth(),
-                )
-                if (isLoggedIn) {
-                    Spacer(Modifier.height(8.dp))
-                    Text(
-                        stringResource(Res.string.summary_lang_capture_identity_note),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                } else {
-                    Spacer(Modifier.height(12.dp))
-                    OutlinedTextField(
-                        value = langEmail,
-                        onValueChange = { langEmail = it; langEmailInvalid = false },
-                        label = { Text(stringResource(Res.string.feedback_email_placeholder)) },
-                        singleLine = true,
-                        isError = langEmailInvalid,
-                        enabled = !langSubmitting,
-                        supportingText = if (langEmailInvalid) {
-                            { Text(stringResource(Res.string.feedback_email_invalid), color = MaterialTheme.colorScheme.error) }
-                        } else null,
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(
-                enabled = langInput.isNotBlank() && !langSubmitting,
-                onClick = {
-                    val lang = langInput.trim()
-                    if (lang.isEmpty()) return@TextButton
-                    val email = langEmail.trim()
-                    if (!isLoggedIn && email.isNotEmpty() && !isValidEmail(email)) {
-                        langEmailInvalid = true
-                        return@TextButton
-                    }
-                    langSubmitting = true
-                    langError = null
-                    scope.launch {
-                        val identityLine = if (isLoggedIn) {
-                            val login = globalSettingsManager.currentGithubLogin()
-                            val userId = globalSettingsManager.currentGithubUserId()
-                            when {
-                                login != null && userId != null -> "GitHub：@${login}（id ${userId}）"
-                                login != null -> "GitHub：@${login}"
-                                else -> "GitHub：已登录（未取到身份）"
-                            }
-                        } else null
-                        val content = buildString {
-                            append("【摘要语言支持请求】期望语言：$lang · 系统语言：${getSystemLanguage()}")
-                            identityLine?.let { append(" · $it") }
-                        }
-                        val submitEmail = if (!isLoggedIn) email.ifEmpty { null } else null
-                        TrendingRepository.shared.submitFeedback(content, submitEmail).fold(
-                            onSuccess = {
-                                langSubmitting = false
-                                onDismiss()
-                                trackEvent("settings_summary_language_sponsor", mapOf("language" to lang))
-                                ProSponsor.openSponsorPage(ProSponsor.SOURCE_SETTINGS_LANGUAGE)
-                            },
-                            onFailure = { e ->
-                                langSubmitting = false
-                                langError = if ((e as? ApiException)?.statusCode == 429) errRate else errGeneric
-                            },
-                        )
-                    }
-                }
-            ) {
-                if (langSubmitting) {
-                    LoadingIndicator(modifier = Modifier.size(24.dp))
-                } else {
-                    Text(stringResource(Res.string.summary_language_sponsor))
-                }
-            }
-        },
-        dismissButton = {
-            TextButton(enabled = !langSubmitting, onClick = onDismiss) {
-                Text(stringResource(Res.string.cancel))
-            }
-        }
     )
 }
